Wrapper for lubridate as.interval

as.interval changes difftime, Duration, Period and numeric class objects to intervals that begin at the specified date-time. Numeric objects are first coerced to timespans equal to the numeric value in seconds.

# S4 method for timespan
as.interval(x, start, ...)

Arguments

x

a duration, difftime, period, or numeric object that describes the length of the interval

start

a POSIXt or Date object that describes when the interval begins

...

additional arguments to pass to as.interval

Value

an interval object

Details

as.interval can be used to create accurate transformations between Period objects, which measure time spans in variable length units, and Duration objects, which measure timespans as an exact number of seconds. A start date- time must be supplied to make the conversion. Lubridate uses this start date to look up how many seconds each variable length unit (e.g. month, year) lasted for during the time span described. See as.duration(), as.period().

See also

Examples

diff <- make_difftime(days = 31) #difftime
#> Error in make_difftime(days = 31): could not find function "make_difftime"
as.interval(diff, ymd("2009-01-01"))
#> Error in .setupMethodsTables(fdef, initialize = TRUE): no slot of name "group" for this object of class "derivedDefaultMethod"
as.interval(diff, ymd("2009-02-01"))
#> Error in .setupMethodsTables(fdef, initialize = TRUE): no slot of name "group" for this object of class "derivedDefaultMethod"

dur <- duration(days = 31) #duration
#> Error in duration(days = 31): could not find function "duration"
as.interval(dur, ymd("2009-01-01"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'as.interval': object 'dur' not found
as.interval(dur, ymd("2009-02-01"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'as.interval': object 'dur' not found

per <- period(months = 1) #period
#> Error in period(months = 1): could not find function "period"
as.interval(per, ymd("2009-01-01"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'as.interval': object 'per' not found
as.interval(per, ymd("2009-02-01"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'as.interval': object 'per' not found

as.interval(3600, ymd("2009-01-01")) #numeric
#> Error in ymd("2009-01-01"): could not find function "ymd"