Datetime component access and assignment functions

Datetime component access and assignment functions

year(x)

year(x) <- value

month(x)

month(x) <- value

mday(x)

mday(x) <- value

day(x)

day(x) <- value

hour(x)

hour(x) <- value

minute(x)

minute(x) <- value

second(x)

second(x) <- value

tz(x)

tz(x) <- value

# S3 method for partial_time
year(x)

\method{year}{partial_time}(x) <- value

# S4 method for partial_time
year(x) <- value

# S3 method for partial_time
month(x)

\method{month}{partial_time}(x) <- value

# S4 method for partial_time
month(x) <- value

# S3 method for partial_time
mday(x)

\method{day}{partial_time}(x) <- value

# S4 method for partial_time
day(x) <- value

# S3 method for partial_time
hour(x)

\method{hour}{partial_time}(x) <- value

# S4 method for partial_time
hour(x) <- value

# S3 method for partial_time
minute(x)

\method{minute}{partial_time}(x) <- value

# S4 method for partial_time
minute(x) <- value

# S3 method for partial_time
second(x)

\method{second}{partial_time}(x) <- value

# S4 method for partial_time
second(x) <- value

# S3 method for partial_time
tz(x)

\method{tz}{partial_time}(x) <- value

Arguments

x

A time-like object to access or assign to

value

For assignment, a value to assign

Note

Care is taken to make these functionas as compatible as possible with similar datetime packages. However, some functions may be masked and cause errors using their masking functions.