Subsetting assignment of partial_time objects

Subsetting assignment of partial_time objects

# S3 method for partial_time
[(x, i, ...) <- value

Arguments

x

object from which to extract element(s) or in which to replace element(s).

i

indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, ... can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, ... can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

...

arguments unused

value

typically an array-like R object of a similar class as x.

Examples

x <- as.parttime(c("2019", "2019-02", "2019-02-02"))
x[c(1, 3)] <- as.parttime(c("2000", "1999"))