Impute a partial time object with a timestamp or specific fields

Impute a partial time object with a timestamp or specific fields

impute_time(x, time, tz, ...)

impute_date(x, time, ..., res = "day")

impute_time_min(x, tz = "-1200", ...)

impute_date_min(x, ..., res = "day")

impute_time_max(x, tz = "+1400", ...)

impute_date_max(x, ..., res = "day")

impute_time_mid(x, tz = "GMT", ...)

impute_date_mid(x, ..., res = "day")

# S3 method for default
impute_time(x, time, tz = "GMT", ...)

# S3 method for POSIXt
impute_time(x, time, tz = "GMT", ...)

# S3 method for partial_time
impute_time(x, time, tz = "GMT", ..., res = NULL)

Arguments

x

a datetime-like object to impute

time

a datetime-like object to use for imputation

tz

a character timezone name for imputation, a character value to use as the timezone part of the datetime or an numeric minute offset.

...

additional individual named fields to impute. Can be one of "year", "month", "day", "hour", "min", "sec", "secfrac", "tzhour", "tzmin".

res

the highest resolution datetime field used for imputation. Either a character value represented the highest resolution field or NULL to impute all fields. For the impute_date family of functions, defaults to "day", or NULL otherwise.

Value

a new partial_time with specified fields imputed